python/requirements.txt
http://xael.org/norman/python/pyclamd/pyClamd-0.3.3.tar.gz
https://github.com/openplanets/fido/archive/1.3.0.tar.gz
python-dateutil